[caffe2] Eliminate implicit calls to strlen when using the RECORD_FUNCTION macros (#153567)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/153567

With the way these were written, any string literals that were being passed in, like `__func__`, did not get inlined deep enough for the conversion to `std::string_view` to still be able to know the length of the string at compile time.

With this new version, it forces the conversion to `std::string_view` to occur in the function that uses the `RECORD_FUNCTION` macro, which means the length of the string literal is available at compile time, and doesn't have to be calculated at runtime.

Test Plan: contbuilds

Differential Revision: D74681042